Derek-CA1/requirements.txt
gpiozero
Adafruit_Python_DHT
twilio
pyserial
Flask
gevent
mysql-connector-python  
mysql-connector